+++ src/main/java/callprotector/spring/domain/chat/dto/request/ChatbotRequestDTO.java
@@ -1,4 +1,4 @@
-package callprotector.spring.web.dto.request;
+package callprotector.spring.domain.chat.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
