+++ src/main/java/callprotector/spring/domain/chat/dto/response/ChatbotResponseDTO.java
@@ -1,4 +1,4 @@
-package callprotector.spring.web.dto.response;
+package callprotector.spring.domain.chat.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
