+++ src/main/java/callprotector/spring/domain/chat/service/ChatSessionServiceImpl.java
@@ -1,9 +1,9 @@
-package callprotector.spring.service.ChatSessionService;
+package callprotector.spring.domain.chat.service;
 
-import callprotector.spring.domain.ChatSession;
+import callprotector.spring.domain.chat.entity.ChatSession;
 import callprotector.spring.domain.user.entity.User;
-import callprotector.spring.repository.ChatSessionRepository;
-import callprotector.spring.web.dto.response.ChatSessionResponseDTO;
+import callprotector.spring.domain.chat.repository.ChatSessionRepository;
+import callprotector.spring.domain.chat.dto.response.ChatSessionResponseDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
